backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
pymongo==4.4.0
python-dotenv==1.0.0
flask-jwt-extended==4.4.4
passlib[bcrypt]==1.7.4
